refactor: decompose generator_teacher.py (2740→429 lines)
Split the monolithic generator_teacher.py into focused modules:

- models.py (51 lines): Exercise, QuizQuestion, Lesson, LessonResult dataclasses
- validators.py (125 lines): 16 exercise validators + VALIDATORS dispatch map
- curriculum/ package (14 files, 122-219 lines each): one module per lesson
- generator_teacher.py (429 lines): GeneratorTeacher class only

Public API preserved — `from amplihack.agents.teaching import GeneratorTeacher`
still works. The __init__.py now also re-exports model classes.

Closes #4362
