+++ src/amplihack/agents/teaching/__init__.py
@@ -1,5 +1,6 @@
 """Teaching agent for the goal-seeking agent generator and eval system."""
 
 from .generator_teacher import GeneratorTeacher
+from .models import Exercise, Lesson, LessonResult, QuizQuestion
 
-__all__ = ["GeneratorTeacher"]
+__all__ = ["GeneratorTeacher", "Exercise", "Lesson", "LessonResult", "QuizQuestion"]
